Make containerized_frames an optional feature
