+++ crates/rqd/Cargo.toml
@@ -6,6 +6,12 @@ version = { workspace = true }
 license = "Apache-2.0"
 description = "Agent for OpenCue"
 
+[features]
+default = []
+# Containerized Frames is a feature that allows rqd to run in a containerized environment.
+# This feature is highly experimental and may not be stable.
+containerized_frames = ["bollard"]
+
 [[bin]]
 path = "src/main.rs"
 name = "openrqd"
@@ -16,6 +22,7 @@ opencue-proto = { path = "../opencue-proto" }
 # dist-lock = { path = "../dist-lock" }
 
 # External Dependencies
+bollard = { version = "0.18.1", optional = true }
 chrono = "0.4.38"
 bincode = "1.3"
 bytesize = { version = "1.2.0", features = ["serde"] }
@@ -53,7 +60,6 @@ pin-project-lite = "0.2.16"
 tower-service = "0.3.3"
 http-body-util = "0.1.3"
 rand = "0.9.1"
-bollard = "0.18.1"
 libc = "0.2"
 
 [dev-dependencies]
