+++ crates/rqd/src/frame/mod.rs
@@ -3,3 +3,6 @@ mod frame_cmd;
 mod logging;
 pub mod manager;
 pub mod running_frame;
+
+#[cfg(feature = "containerized_frames")]
+pub mod docker_running_frame;
