+++ crates/rqd/src/frame/running_frame.rs
@@ -6,33 +6,19 @@ use std::os::fd::{FromRawFd, RawFd};
 use std::os::unix::process::ExitStatusExt;
 use std::time::SystemTime;
 use std::{
-    cmp,
     collections::HashMap,
     env,
     fmt::Display,
     path::Path,
     process::ExitStatus,
-    str::FromStr,
     sync::{Arc, RwLock},
 };
 use std::{process::Stdio, thread};
 use tokio::time::{self, Duration};
 
-use bollard::{
-    Docker,
-    container::{
-        self, AttachContainerOptions, AttachContainerResults, CreateContainerOptions,
-        StartContainerOptions, WaitContainerOptions,
-    },
-    secret::{
-        ContainerWaitResponse, DeviceMapping, HostConfig, Mount, MountBindOptions,
-        MountBindOptionsPropagationEnum, MountTypeEnum,
-    },
-};
 use bytesize::KIB;
 use chrono::{DateTime, Local};
-use futures::StreamExt;
-use itertools::{Either, Itertools};
+use itertools::Itertools;
 use tokio::fs::File;
 use tokio::io::{AsyncReadExt, AsyncWriteExt, BufWriter};
 use tokio::{io::AsyncBufReadExt, task::JoinHandle};
@@ -59,12 +45,12 @@ pub struct RunningFrame {
     pub layer_id: Uuid,
     frame_stats: RwLock<ProcessStats>,
     pub log_path: String,
-    uid: u32,
-    gid: u32,
-    config: RunnerConfig,
+    pub(super) uid: u32,
+    pub(super) gid: u32,
+    pub(super) config: RunnerConfig,
     pub thread_ids: Option<Vec<u32>>,
     pub gpu_list: Option<Vec<u32>>,
-    env_vars: HashMap<String, String>,
+    pub(super) env_vars: HashMap<String, String>,
     pub hostname: String,
     raw_stdout_path: String,
     raw_stderr_path: String,
@@ -315,7 +301,7 @@ impl RunningFrame {
     /// If the frame has already started or finished, log the error and don't change the status.
     /// Returning an error is pointless as we want the frame that trigger this transition to finish
     /// regardless
-    fn start(&self, pid: u32) {
+    pub(super) fn start(&self, pid: u32) {
         let mut state = self.state.write().unwrap_or_else(|err| err.into_inner());
 
         match &mut *state {
@@ -438,80 +424,6 @@ impl RunningFrame {
         };
     }
 
-    /// Runs the frame as a docker container.
-    ///
-    /// This method is the main entry point for executing a frame. It:
-    /// 1. Creates a logger for the frame
-    /// 2. Runs the frame command on a new process
-    /// 3. Updates the frame's exit code based on the result
-    /// 4. Cleans up any snapshots created during execution
-    ///
-    /// If the process fails to spawn, it logs the error but doesn't set an exit code.
-    /// The method handles both successful and failed execution scenarios.
-    pub async fn run_docker(&self, recover_mode: bool) {
-        let logger_base = FrameLoggerBuilder::from_logger_config(
-            self.log_path.clone(),
-            &self.config,
-            self.config.run_as_user.then(|| (self.uid, self.gid)),
-        );
-        if let Err(err) = logger_base {
-            error!("Failed to create log stream for {}: {}", self.log_path, err);
-            if let Err(err) = self.fail_before_start() {
-                error!("Failed to mark frame {} as finished. {}", self, err);
-            }
-            return;
-        }
-        let logger = Arc::new(logger_base.unwrap());
-
-        let exit_code = if recover_mode {
-            match self.recover_inner(Arc::clone(&logger)).await {
-                Ok((exit_code, exit_signal)) => {
-                    if let Err(err) = self.finish(exit_code, exit_signal) {
-                        warn!("Failed to mark frame {} as finished. {}", self, err);
-                    }
-                    logger.writeln(&self.write_footer());
-                    Some(exit_code)
-                }
-                Err(err) => {
-                    let msg = format!("Frame {} failed to be recovered. {}", self.to_string(), err);
-                    logger.writeln(&msg);
-                    error!(msg);
-                    None
-                }
-            }
-        } else {
-            let run_result = self.run_docker_inner(Arc::clone(&logger)).await;
-            match run_result {
-                Ok((exit_code, exit_signal)) => {
-                    if let Err(err) = self.finish(exit_code, exit_signal) {
-                        warn!("Failed to mark frame {} as finished. {}", self, err);
-                    }
-                    logger.writeln(&self.write_footer());
-                    Some(exit_code)
-                }
-                Err(err) => {
-                    let msg = format!("Frame {} failed to be spawned. {}", self.to_string(), err);
-                    logger.writeln(&msg);
-                    error!(msg);
-                    if let Err(err) = self.fail_before_start() {
-                        error!("Failed to mark frame {} as finished. {}", self, err);
-                    }
-                    None
-                }
-            }
-        };
-        if let Err(err) = self.clear_snapshot().await {
-            // Only warn if a job was actually launched
-            if exit_code.is_some() {
-                warn!(
-                    "Failed to clear snapshot {}: {}",
-                    self.snapshot_path().unwrap_or("empty_path".to_string()),
-                    err
-                );
-            }
-        };
-    }
-
     #[cfg(any(target_os = "linux", target_os = "macos"))]
     async fn run_inner(&self, logger: FrameLogger) -> Result<(i32, Option<i32>)> {
         use nix::libc;
@@ -594,7 +506,7 @@ impl RunningFrame {
             .into_diagnostic()
             .wrap_err(format!("Command for {self} didn't start!"))?;
 
-        let (exit_code, exit_signal) = Self::interprete_output(Either::Left(output));
+        let (exit_code, exit_signal) = Self::interprete_output(output);
 
         let msg = match exit_code {
             0 => format!("Frame {}(pid={}) finished successfully", self, pid),
@@ -612,19 +524,9 @@ impl RunningFrame {
         Ok((exit_code, exit_signal))
     }
 
-    fn interprete_output(output: Either<ExitStatus, ContainerWaitResponse>) -> (i32, Option<i32>) {
-        let (mut exit_signal, mut exit_code) = match output {
-            Either::Left(exit_status) => {
-                let exit_signal = exit_status.signal();
-                let exit_code = exit_status.code().unwrap_or(1);
-                (exit_signal, exit_code)
-            }
-            Either::Right(wait_response) => {
-                let exit_signal = None;
-                let exit_code = wait_response.status_code as i32;
-                (exit_signal, exit_code)
-            }
-        };
+    fn interprete_output(exit_status: ExitStatus) -> (i32, Option<i32>) {
+        let mut exit_signal = exit_status.signal();
+        let mut exit_code = exit_status.code().unwrap_or(1);
 
         // If the cmd wrapper interprets the signal as an output, 128 needs to be subtracted
         // from the code to recover the received signal
@@ -633,201 +535,6 @@ impl RunningFrame {
             exit_signal = Some(exit_code - 128);
         }
         (exit_code, exit_signal)
-    }
-
-    pub async fn run_docker_inner(&self, logger: FrameLogger) -> Result<(i32, Option<i32>)> {
-        logger.writeln(self.write_header().as_str());
-
-        let docker = Docker::connect_with_socket_defaults()
-            .into_diagnostic()
-            .wrap_err("Failed to connect to docker socket.")?;
-
-        let image = Some(self.config.get_docker_image(&self.request.os));
-        let env: Option<Vec<String>> = Some(
-            self.env_vars
-                .iter()
-                .map(|(k, v)| format!("{k}={v}"))
-                .collect(),
-        );
-        let working_dir = Some(self.config.temp_path.clone());
-
-        // Build Command
-        let mut command =
-            FrameCmdBuilder::new(&self.config.shell_path, self.entrypoint_file_path.clone());
-        if self.config.desktop_mode {
-            command.with_nice();
-        }
-        if let Some(cpu_list) = &self.thread_ids {
-            command.with_taskset(cpu_list.clone());
-        }
-
-        command.with_become_user(self.uid, self.gid, self.request.user_name.clone());
-        let (_cmd, cmd_str) = command
-            .with_frame_cmd(self.request.command.clone())
-            .with_exit_file(self.exit_file_path.clone())
-            .build()?;
-        let entrypoint = Some(vec![
-            // Execute entrypoint file
-            self.entrypoint_file_path.clone(),
-        ]);
-
-        trace!("Running {}: {}", self.entrypoint_file_path, cmd_str);
-        logger.writeln(format!("Running {}:", self.entrypoint_file_path).as_str());
-
-        let host_config = self.build_docker_host_config();
-        let container_name = format!(
-            "frame_{}_{}",
-            self.request.job_name, self.request.resource_id
-        );
-
-        let _container = &docker
-            .create_container(
-                Some(CreateContainerOptions {
-                    name: container_name.clone(),
-                    platform: None,
-                }),
-                container::Config::<String> {
-                    hostname: Some(self.hostname.clone()),
-                    attach_stdout: Some(true),
-                    attach_stderr: Some(true),
-                    tty: Some(true),
-                    env,
-                    image,
-                    working_dir,
-                    entrypoint,
-                    host_config: Some(host_config),
-                    ..Default::default()
-                },
-            )
-            .await
-            .into_diagnostic()
-            .wrap_err("Failed to create container")?;
-
-        let _ = &docker
-            .start_container(&container_name, None::<StartContainerOptions<String>>)
-            .await
-            .into_diagnostic()
-            .wrap_err("Failed to start container")?;
-
-        let AttachContainerResults {
-            output: mut log_stream,
-            input: _,
-        } = docker
-            .attach_container(
-                &container_name,
-                Some(AttachContainerOptions::<String> {
-                    stdin: Some(false),
-                    stdout: Some(true),
-                    stderr: Some(true),
-                    stream: Some(true),
-                    logs: Some(true),
-                    detach_keys: Some("ctrl-c".to_string()),
-                }),
-            )
-            .await
-            .into_diagnostic()?;
-        // Read the pid from from first line of the log
-        let first_line = log_stream
-            .next()
-            .await
-            .ok_or_else(|| miette!("Failed to attach to log stream"))?;
-
-        let pid_result = first_line
-            .into_diagnostic()
-            .wrap_err("Failed to read pid from log")
-            .and_then(|line| line.to_string().parse::<u32>().into_diagnostic());
-
-        let pid = match pid_result {
-            Ok(pid) => pid,
-            Err(err) => {
-                // Clean up container before returning the error
-                let _ = docker.remove_container(&container_name, None).await;
-                return Err(err);
-            }
-        };
-
-        // Update frame state with frame pid
-        self.start(pid);
-
-        info!(
-            "Frame {self} started with pid {pid}, with taskset {}",
-            self.taskset()
-        );
-
-        let _ = self.create_snapshot().await;
-        let log_watcher_handle = tokio::task::spawn(async move {
-            while let Some(Ok(output)) = log_stream.next().await {
-                logger.write(output.into_bytes().as_ref());
-            }
-        });
-
-        let output_stream =
-            docker.wait_container(&container_name, None::<WaitContainerOptions<String>>);
-        let mut exit_code = 1;
-        let mut exit_signal = None;
-        if let Some(Ok(output)) = output_stream.take(1).next().await {
-            (exit_code, exit_signal) = Self::interprete_output(Either::Right(output));
-        }
-        log_watcher_handle.abort();
-
-        let msg = match exit_code {
-            0 => format!("Frame {}(pid={}) finished successfully", self, pid),
-            _ => format!(
-                "Frame {}(pid={}) finished with exit_code={} and exit_signal={}. Log: {}",
-                self,
-                pid,
-                exit_code,
-                exit_signal.unwrap_or(0),
-                self.log_path,
-            ),
-        };
-        info!(msg);
-
-        Ok((exit_code, exit_signal))
-    }
-
-    pub fn build_docker_host_config(&self) -> HostConfig {
-        let mounts = Some(
-            self.config
-                .docker_mounts
-                .iter()
-                .map(|mount| Mount {
-                    target: Some(mount.target.clone()),
-                    source: Some(mount.source.clone()),
-                    typ: Some(
-                        MountTypeEnum::from_str(mount.typ.as_str()).unwrap_or(MountTypeEnum::BIND),
-                    ),
-                    bind_options: Some(MountBindOptions {
-                        propagation: Some(
-                            MountBindOptionsPropagationEnum::from_str(
-                                &mount.bind_propagation.as_str(),
-                            )
-                            .unwrap_or(MountBindOptionsPropagationEnum::SLAVE),
-                        ),
-                        ..Default::default()
-                    }),
-
-                    ..Default::default()
-                })
-                .collect(),
-        );
-
-        // Docker requires memory limits higer than 6MG
-        let soft_memory_limit = cmp::max(self.request.soft_memory_limit, 6291456) * 1000;
-        let hard_memory_limit = cmp::max(self.request.hard_memory_limit, 6291456) * 1000;
-        HostConfig {
-            devices: Some(vec![DeviceMapping {
-                path_on_host: Some("/dev/fuse".to_string()),
-                path_in_container: Some("/dev/fuse".to_string()),
-                cgroup_permissions: None,
-            }]),
-            auto_remove: Some(true),
-            mounts,
-            privileged: Some(true),
-            memory_reservation: Some(soft_memory_limit),
-            memory: Some(hard_memory_limit),
-            ..Default::default()
-        }
     }
 
     #[cfg(target_os = "windows")]
@@ -880,7 +587,7 @@ impl RunningFrame {
     /// # Errors
     /// Returns an error if the frame doesn't have a valid PID or if process monitoring fails
     #[cfg(any(target_os = "linux", target_os = "macos"))]
-    async fn recover_inner(&self, logger: FrameLogger) -> Result<(i32, Option<i32>)> {
+    pub(super) async fn recover_inner(&self, logger: FrameLogger) -> Result<(i32, Option<i32>)> {
         logger.writeln(self.write_header().as_str());
 
         let pid = self.pid().ok_or(miette!(
@@ -1196,7 +903,7 @@ impl RunningFrame {
         Ok(current_position)
     }
 
-    fn snapshot_path(&self) -> Result<String> {
+    pub(super) fn snapshot_path(&self) -> Result<String> {
         let pid = self
             .pid()
             .ok_or_else(|| miette!("No pid available for frame snapshot"))?;
@@ -1210,7 +917,7 @@ impl RunningFrame {
     /// Save a snapshot of the frame into disk to enable recovering its status in case
     /// rqd restarts.
     ///
-    async fn create_snapshot(&self) -> Result<()> {
+    pub(super) async fn create_snapshot(&self) -> Result<()> {
         let snapshot_path = self.snapshot_path()?;
         let file = File::create(&snapshot_path).await.into_diagnostic()?;
         let mut writer = BufWriter::new(file);
@@ -1226,7 +933,7 @@ impl RunningFrame {
         Ok(())
     }
 
-    async fn clear_snapshot(&self) -> Result<()> {
+    pub(super) async fn clear_snapshot(&self) -> Result<()> {
         let snapshot_path = self.snapshot_path()?;
         tokio::fs::remove_file(snapshot_path)
             .await
@@ -1300,7 +1007,7 @@ impl RunningFrame {
         system.process(Pid::from_u32(pid)).is_some()
     }
 
-    fn write_header(&self) -> String {
+    pub(super) fn write_header(&self) -> String {
         let env_var_list = self
             .env_vars
             .iter()
@@ -1350,7 +1057,7 @@ Environment Variables:
         )
     }
 
-    fn write_footer(&self) -> String {
+    pub(super) fn write_footer(&self) -> String {
         let frame_stats_lock = self
             .frame_stats
             .read()
@@ -1435,7 +1142,7 @@ Render Frame Completed
         }
     }
 
-    fn taskset(&self) -> String {
+    pub(super) fn taskset(&self) -> String {
         self.thread_ids
             .clone()
             .unwrap_or(vec![0])
